2 new levels + bird entity + fish entity complete
Designed 2 levels: level 3 sees the introduction of the fish and level 4 has a bird. Bird fires a projectile every turn towards the snail if it is in range. Fish just jumps up and down from a water tile until it finds a wall, stays in the tile below the wall for 1 turn and goes back down.

There is a really small visual bug with the fish when it is in the water which I don't really know how to fix as of yet. I believe it is due to the water sprite moving just in front of the fish and then behind it over time. This does not affect gameplay so I think we are good to merge if this is the only problem with this PR. I will probably spend the remaining time fixing and finding bugs within the gameplay since my part is practically done, so that will be my focus from now on. Let me know if you have any questions or concerns over this PR .